+++ packages/react/src/components/Card/CardContent/index.ts
@@ -1,1 +1,1 @@
-export * from './CardContent';
+export * from '../CardContent';
