+++ packages/react/src/components/Card/CardFooter/index.ts
@@ -1,1 +1,1 @@
-export * from './CardFooter';
+export * from '../CardFooter';
