+++ packages/react/src/components/Card/CardGroup/index.ts
@@ -1,1 +1,1 @@
-export * from './CardGroup';
+export * from '../CardGroup';
