+++ packages/react/src/components/Card/CardHeader/index.ts
@@ -1,1 +1,1 @@
-export * from './CardHeader';
+export * from '../CardHeader';
